#BUILD: edits to build system and restructuring of linalg module
