+++ modules/linalg/arm/mtx_arm_vec_i32.cpp
@@ -30,7 +30,7 @@
  * WARRANTY OF ANY KIND, either express or implied.
  *
  ************************************************************************/
-#include "../../include/linalg/mtx.hpp"
+#include "../../../include/linalg/mtx.hpp"
 #include <cassert>
 #include <cstddef>
 #include <cstdint>
