+++ modules/linalg/avx/mtx_avx2_vec_f64.cpp
@@ -30,7 +30,7 @@
  * WARRANTY OF ANY KIND, either express or implied.
  *
  ************************************************************************/
-#include "../../include/linalg/mtx.hpp"
+#include "../../../include/linalg/mtx.hpp"
 #include <cassert>
 #include <cstddef>
 #include <cstdint>
@@ -41,46 +41,50 @@
 
 /************************************************************************
  *
- * Matrix Operations for SSE ISA
+ * Matrix Operations for AVX ISA
  *
  ************************************************************************/
-#elif defined(__SSE2__)
-// SSE2
-#include <emmintrin.h>
-#include <smmintrin.h>
+#if defined(__AVX2__)
+
+// AVX family intrinsics
+#include <immintrin.h>
 
 /************************************************************************
  *
  * Matrix Operations on vector<vector>
  *
  ************************************************************************/
-
+// matrix addition using Intel intrinsics, accepts type double
 void gpmp::linalg::Mtx::mtx_add(const std::vector<std::vector<double>> &A,
                                 const std::vector<std::vector<double>> &B,
                                 std::vector<std::vector<double>> &C) {
     const int rows = A.size();
     const int cols = A[0].size();
 
-    for (int i = 0; i < rows; ++i) {
-        int j = 0;
-        // requires at least size 2x2 matrices for SSE2
-        for (; j < cols - 1; j += 2) {
-            // load 2 elements from A, B, and C matrices using SSE2
-            __m128d a = _mm_loadu_pd(&A[i][j]);
-            __m128d b = _mm_loadu_pd(&B[i][j]);
-            __m128d c = _mm_loadu_pd(&C[i][j]);
+    if (rows > 8) {
+        for (int i = 0; i < rows; ++i) {
+            int j = 0;
+            // requires at least size 4x4 matrices
+            for (; j < cols - 3; j += 4) {
+                // load 4 elements from A, B, and C matrices using SIMD
+                __m256d a = _mm256_loadu_pd(&A[i][j]);
+                __m256d b = _mm256_loadu_pd(&B[i][j]);
+                __m256d c = _mm256_loadu_pd(&C[i][j]);
 
-            // perform vectorized addition
-            c = _mm_add_pd(a, b);
+                // perform vectorized addition
+                c = _mm256_add_pd(a, b);
 
-            // store the result back to the C matrix
-            _mm_storeu_pd(&C[i][j], c);
+                // store the result back to the C matrix
+                _mm256_storeu_pd(&C[i][j], c);
+            }
+
+            // handle the remaining elements that are not multiples of 4
+            for (; j < cols; ++j) {
+                C[i][j] = A[i][j] + B[i][j];
+            }
         }
-
-        // handle the remaining elements that are not multiples of 2
-        for (; j < cols; ++j) {
-            C[i][j] = A[i][j] + B[i][j];
-        }
+    } else {
+        std_mtx_add(A, B, C);
     }
 }
 
@@ -93,21 +97,21 @@ void gpmp::linalg::Mtx::mtx_sub(const std::vector<std::vector<double>> &A,
 
     for (int i = 0; i < rows; ++i) {
         int j = 0;
-        // requires at least size 2x2 matrices for SSE2
-        for (; j < cols - 1; j += 2) {
-            // load 2 elements from A, B, and C matrices using SSE2
-            __m128d a = _mm_loadu_pd(&A[i][j]);
-            __m128d b = _mm_loadu_pd(&B[i][j]);
-            __m128d c = _mm_loadu_pd(&C[i][j]);
+        // requires at least size 4x4 matrices
+        for (; j < cols - 3; j += 4) {
+            // load 4 elements from A, B, and C matrices using SIMD
+            __m256d a = _mm256_loadu_pd(&A[i][j]);
+            __m256d b = _mm256_loadu_pd(&B[i][j]);
+            __m256d c = _mm256_loadu_pd(&C[i][j]);
 
             // perform vectorized subtraction
-            c = _mm_sub_pd(a, b);
+            c = _mm256_sub_pd(a, b);
 
             // store the result back to the C matrix
-            _mm_storeu_pd(&C[i][j], c);
+            _mm256_storeu_pd(&C[i][j], c);
         }
 
-        // handle the remaining elements that are not multiples of 2
+        // handle the remaining elements that are not multiples of 4
         for (; j < cols; ++j) {
             C[i][j] = A[i][j] - B[i][j];
         }
@@ -123,28 +127,28 @@ void gpmp::linalg::Mtx::mtx_mult(const std::vector<std::vector<double>> &A,
     const int cols_b = B[0].size();
 
     for (int i = 0; i < rows_a; ++i) {
-        for (int j = 0; j < cols_b; j += 2) {
+        for (int j = 0; j < cols_b; j += 4) {
             // initialize a vector of zeros for the result
-            __m128d c = _mm_setzero_pd();
+            __m256d c = _mm256_setzero_pd();
 
             for (int k = 0; k < cols_a; ++k) {
-                // load 2 elements from matrices A and B using SSE2
-                __m128d a = _mm_set1_pd(A[i][k]);
-                __m128d b = _mm_loadu_pd(&B[k][j]);
+                // load 4 elements from matrices A and B using SIMD
+                __m256d a = _mm256_set1_pd(A[i][k]);
+                __m256d b = _mm256_loadu_pd(&B[k][j]);
 
                 // perform vectorized multiplication
-                __m128d prod = _mm_mul_pd(a, b);
+                __m256d prod = _mm256_mul_pd(a, b);
 
                 // perform vectorized addition
-                c = _mm_add_pd(c, prod);
+                c = _mm256_add_pd(c, prod);
             }
 
             // store the result back to the C matrix
-            _mm_storeu_pd(&C[i][j], c);
+            _mm256_storeu_pd(&C[i][j], c);
         }
 
-        // handle the remaining elements that are not multiples of 2
-        for (int j = cols_b - cols_b % 2; j < cols_b; ++j) {
+        // handle the remaining elements that are not multiples of 4
+        for (int j = cols_b - cols_b % 4; j < cols_b; ++j) {
             double sum = 0.0;
 
             for (int k = 0; k < cols_a; ++k) {
@@ -156,23 +160,35 @@ void gpmp::linalg::Mtx::mtx_mult(const std::vector<std::vector<double>> &A,
     }
 }
 
-// transpose matrices of type double using Intel intrinsics
 void gpmp::linalg::Mtx::mtx_tpose(std::vector<std::vector<double>> &matrix) {
     const int rows = matrix.size();
     const int cols = matrix[0].size();
 
-    for (int i = 0; i < rows; i += 2) {
-        for (int j = i; j < cols; j += 2) {
-            __m128d row1 = _mm_loadu_pd(&matrix[i][j]);
-            __m128d row2 = _mm_loadu_pd(&matrix[i + 1][j]);
+    for (int i = 0; i < rows; i += 4) {
+        for (int j = i; j < cols; j += 4) {
+            __m256d row1 = _mm256_loadu_pd(&matrix[i][j]);
+            __m256d row2 = _mm256_loadu_pd(&matrix[i + 1][j]);
+            __m256d row3 = _mm256_loadu_pd(&matrix[i + 2][j]);
+            __m256d row4 = _mm256_loadu_pd(&matrix[i + 3][j]);
 
-            // transpose 2x2 submatrix
-            __m128d tmp1 = _mm_unpacklo_pd(row1, row2);
-            __m128d tmp2 = _mm_unpackhi_pd(row1, row2);
+            __m256d tmp1, tmp2, tmp3, tmp4;
 
-            // store the transposed 2x2 submatrix back to the matrix
-            _mm_storeu_pd(&matrix[i][j], tmp1);
-            _mm_storeu_pd(&matrix[i + 1][j], tmp2);
+            // Transpose 4x4 submatrix
+            tmp1 = _mm256_unpacklo_pd(row1, row2);
+            tmp2 = _mm256_unpackhi_pd(row1, row2);
+            tmp3 = _mm256_unpacklo_pd(row3, row4);
+            tmp4 = _mm256_unpackhi_pd(row3, row4);
+
+            row1 = _mm256_permute2f128_pd(tmp1, tmp3, 0x20);
+            row2 = _mm256_permute2f128_pd(tmp2, tmp4, 0x20);
+            row3 = _mm256_permute2f128_pd(tmp1, tmp3, 0x31);
+            row4 = _mm256_permute2f128_pd(tmp2, tmp4, 0x31);
+
+            // Store the transposed 4x4 submatrix back to the matrix
+            _mm256_storeu_pd(&matrix[i][j], row1);
+            _mm256_storeu_pd(&matrix[i + 1][j], row2);
+            _mm256_storeu_pd(&matrix[i + 2][j], row3);
+            _mm256_storeu_pd(&matrix[i + 3][j], row4);
         }
     }
 }
