+++ modules/linalg/avx/vector_avx2_i32.cpp
@@ -31,7 +31,7 @@
  *
  ************************************************************************/
 
-#include "../../include/linalg/vector.hpp"
+#include "../../../include/linalg/vector.hpp"
 #include <cmath>
 #include <cstdint>
 #include <iostream>
